LibCerf

Documentation for LibCerf.

LibCerf.LibCerf — Module

LibCerf

A Julia wrapper for the libcerf library.

libcerf, a self-contained numeric library that provides an efficient and accurate implementation of complex error functions, along with Dawson, Faddeeva, and Voigt functions.

Exported functions

erf(z): Complex error function

erfc(z): Complementary error function

erfcx(z), erfcx(x): Underflow-compensated complementary error function

erfi(z), erfi(x): Imaginary error function

faddeeva_w(z): Faddeeva's scaled complex error function

im_w(x): imaginary part of faddeeva_w(complex(x, 0.0))

im_w(re, im): imaginary part of faddeeva_w(complex(re, im))

re_w(re, im): real part of faddeeva_w(complex(re, im))

dawson(z), dawson(x): Dawson's integral

voigt(x, sigma, gamma): Voigt's convolution of a Gaussian and a Lorentzian

voigt_hwhm(sigma, gamma): Half width at half maximum of the Voigt profile

References

S. G. Johnson, J. Wuttke: libcerf, numeric library for complex error functions, https://jugit.fz-juelich.de/mlz/libcerf

Index

LibCerf.LibCerf

LibCerf.dawson

LibCerf.erf

LibCerf.erfc

LibCerf.erfcx

LibCerf.erfi

LibCerf.faddeeva_w

LibCerf.im_w

LibCerf.re_w

LibCerf.voigt

LibCerf.voigt_hwhm

Error Functions

LibCerf.erf — Function

erf(z)

Error function of complex arguments.

\[\text{erf}(z) = \frac{2}{\sqrt{\pi}} \int_0^z e^{-t^2} dt\]

Examples

julia> [ (x=x, erf=Float32(erf(complex(x)))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, erf::Float32}}:
 (x = -2.0, erf = -0.9953223)
 (x = -1.5, erf = -0.96610516)
 (x = -1.0, erf = -0.8427008)
 (x = -0.5, erf = -0.5204999)
 (x = 0.0, erf = 0.0)
 (x = 0.5, erf = 0.5204999)
 (x = 1.0, erf = 0.8427008)
 (x = 1.5, erf = 0.96610516)
 (x = 2.0, erf = 0.9953223)

julia> [ (x=x, erf=erf(complex(x))) for x in (-Inf, -0.0, 0.0, Inf) ]
4-element Vector{@NamedTuple{x::Float64, erf::ComplexF64}}:
 (x = -Inf, erf = -1.0 + 0.0im)
 (x = -0.0, erf = -0.0 + 0.0im)
 (x = 0.0, erf = 0.0 + 0.0im)
 (x = Inf, erf = 1.0 + 0.0im)

julia> erf(complex(3.14)) + erfc(complex(3.14))
1.0 + 0.0im

See also: erfc(z), erfcx(z), erfi(z)

References

DLMF: 7.2.1

Erf – Wolfram MathWorld

Real part plot of erf(z):

using UnicodePlots
re_erf(x) = real(LibCerf.erf(complex(x)))
lineplot(-4, 4, re_erf; title="Re[Erf(z)]", ylabel="erf(x)")

             ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[Erf(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀          
             ┌────────────────────────────────────────┐          
           1 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⣀⠔⠊⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉│ re_erf(x)
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⢀⡔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⡜⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⡸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⢠⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣇⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡟⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
   erf(x)    │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⢤⡧⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡎⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡸⠁⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠇⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡸⠁⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠁⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
          -1 │⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⡠⠔⠋⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│          
             └────────────────────────────────────────┘          
             ⠀-4⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀4⠀          
             ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀          

LibCerf.erfc — Function

erfc(z)

Complementary error function of complex arguments.

\[\text{erfc}(z) = \frac{2}{\sqrt{\pi}} \int_z^\infty e^{-t^2} dt = 1 - \text{erf}(z)\]

Examples

julia> [ (x=x, erfc=Float32(erfc(complex(x)))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, erfc::Float32}}:
 (x = -2.0, erfc = 1.9953222)
 (x = -1.5, erfc = 1.9661051)
 (x = -1.0, erfc = 1.8427008)
 (x = -0.5, erfc = 1.5204998)
 (x = 0.0, erfc = 1.0)
 (x = 0.5, erfc = 0.47950011)
 (x = 1.0, erfc = 0.1572992)
 (x = 1.5, erfc = 0.03389485)
 (x = 2.0, erfc = 0.004677735)

julia> [ (x=x, erfc=erfc(complex(x))) for x in (-Inf, 0.0, Inf) ]
3-element Vector{@NamedTuple{x::Float64, erfc::ComplexF64}}:
 (x = -Inf, erfc = 2.0 - 0.0im)
 (x = 0.0, erfc = 1.0 - 0.0im)
 (x = Inf, erfc = 0.0 - 0.0im)

julia> erf(complex(3.141)) + erfc(complex(3.141))
1.0 + 0.0im

See also: erf(z), erfcx(z)

References

DLMF: 7.2.2

Erfc – Wolfram MathWorld

Real part plot of erfc(z):

using UnicodePlots
re_erfc(x) = real(LibCerf.erfc(complex(x)))
plt = lineplot(-4, 4, re_erfc; title="Re[Erfc(z)]", ylabel="erfc(x)");
lineplot!(plt, 1, 0; name="y = 1")

             ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[Erfc(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           
             ┌────────────────────────────────────────┐           
           2 │⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠑⠢⣄⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_erfc(x)
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢣⡀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ y = 1     
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢱⡀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢣⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⡆⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢱⡀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢇⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
   erfc(x)   │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠼⡧⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣧⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡏⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⢣⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
             │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠈⠣⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           0 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠉⠢⢄⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀│           
             └────────────────────────────────────────┘           
             ⠀-4⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀4⠀           
             ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           

erf(x) and erfc(x) relations:

lineplot(-4, 4, [re_erf, re_erfc]; title="Erf(z) + Erfc(z) = 1", ylim=(2, -1))

           ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Erf(z) + Erfc(z) = 1⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           
           ┌────────────────────────────────────────┐           
         2 │⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠓⠦⡀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_erf(x) 
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⢢⡀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_erfc(x)
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢣⡀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢱⡀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢣⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣧⠀⠀⠀⢀⡤⠒⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⢣⠀⡔⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
   f(x)    │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⢈⣿⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⡜⠀⠣⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡟⠀⠀⠀⠈⠓⠤⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀│           
           │⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⡝⡏⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉⠉│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡸⠁⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠁⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠜⠁⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
        -1 │⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⣀⡤⠖⠁⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           └────────────────────────────────────────┘           
           ⠀-4⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀4⠀           
           ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           

LibCerf.erfcx — Function

erfcx(x)
erfcx(z)

Underflow-compensated complementary error function of real or complex arguments.

\[\text{erfcx}(z) = \exp(z^2) \text{erfc}(z) = w(iz)\]

Examples

julia> [ (x=x, erfcx=Float32(erfcx(complex(x)))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, erfcx::Float32}}:
 (x = -2.0, erfcx = 108.9409)
 (x = -1.5, erfcx = 18.653887)
 (x = -1.0, erfcx = 5.0089803)
 (x = -0.5, erfcx = 1.9523605)
 (x = 0.0, erfcx = 1.0)
 (x = 0.5, erfcx = 0.61569035)
 (x = 1.0, erfcx = 0.42758358)
 (x = 1.5, erfcx = 0.32158542)
 (x = 2.0, erfcx = 0.25539568)

julia> [ (x=x, erfcx=erfcx(complex(x))) for x in (-Inf, 0.0, Inf) ]
3-element Vector{@NamedTuple{x::Float64, erfcx::ComplexF64}}:
 (x = -Inf, erfcx = Inf - 0.0im)
 (x = 0.0, erfcx = 1.0 - 0.0im)
 (x = Inf, erfcx = 0.0 - 0.0im)

julia> erfcx(complex(pi)) - exp(pi^2)*erfc(complex(pi))
0.0 + 0.0im

julia> erfcx(complex(pi)) - faddeeva_w(im * complex(pi))
0.0 - 0.0im

See also: erf(z), erfc(z), faddeeva_w(z)

References

Faddeeva Package

Real part plot of erfcx(z):

using UnicodePlots
re_erfcx(x) = real(LibCerf.erfcx(complex(x)))
lineplot(-1.5, 2, re_erfcx; title="Re[Erfcx(z)]", ylabel="erfcx(x)")

               ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[Erfcx(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀            
               ┌────────────────────────────────────────┐            
            20 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_erfcx(x)
               │⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
   erfcx(x)    │⠀⠀⠀⠀⠀⠀⠈⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⠀⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⠀⠀⢧⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⠀⠀⠈⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⢢⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⢄⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠲⠤⣀⣀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
             0 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⡗⠒⠒⠒⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⣄⣀⣀⣀⣀⣀│            
               └────────────────────────────────────────┘            
               ⠀-2⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀2⠀            
               ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀            

LibCerf.erfi — Function

erfi(x)
erfi(z)

Imaginary error function of real or complex arguments.

\[\text{erfi}(z) = -i \text{erf}(iz)\]

Examples

julia> [ (x=x, erfi=Float32(erfi(complex(x)))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, erfi::Float32}}:
 (x = -2.0, erfi = -18.564802)
 (x = -1.5, erfi = -4.5847335)
 (x = -1.0, erfi = -1.6504258)
 (x = -0.5, erfi = -0.6149521)
 (x = 0.0, erfi = 0.0)
 (x = 0.5, erfi = 0.6149521)
 (x = 1.0, erfi = 1.6504258)
 (x = 1.5, erfi = 4.5847335)
 (x = 2.0, erfi = 18.564802)

julia> [ (x=x, erfi=erfi(complex(x))) for x in (-Inf, -0.0, 0.0, Inf) ]
4-element Vector{@NamedTuple{x::Float64, erfi::ComplexF64}}:
 (x = -Inf, erfi = -Inf + 0.0im)
 (x = -0.0, erfi = -0.0 + 0.0im)
 (x = 0.0, erfi = 0.0 + 0.0im)
 (x = Inf, erfi = Inf + 0.0im)

julia> erfi(complex(pi)) + im*erf(im*complex(pi))
0.0 + 0.0im

See also: erf(z)

References

Erfi – Wolfram MathWorld

Imaginary error function - Wikipedia

Real part plot of erfi(z):

using UnicodePlots
re_erfi(x) = real(LibCerf.erfi(complex(x)))
lineplot(-2, 2, re_erfi; title="Re[Erfi(z)]", ylabel="erfi(x)")

               ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[Erfi(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           
               ┌────────────────────────────────────────┐           
            20 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀│ re_erfi(x)
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜│           
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃│           
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡎⠀│           
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠎⠀⠀│           
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡔⠃⠀⠀⠀│           
               │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⣀⣀⡤⠔⠊⠁⠀⠀⠀⠀⠀│           
   erfi(x)     │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⣤⣤⣤⡤⠤⠤⠤⠤⠤⡷⠶⠶⠶⠶⠮⠭⠭⠭⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│           
               │⠀⠀⠀⠀⠀⢀⡠⠔⠒⠉⠉⠁⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
               │⠀⠀⠀⢠⠔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
               │⠀⠀⡰⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
               │⠀⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
               │⢠⠇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
               │⡎⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
           -20 │⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
               └────────────────────────────────────────┘           
               ⠀-2⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀2⠀           
               ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           

Faddeeva Functions

LibCerf.faddeeva_w — Function

faddeeva_w(z)

Faddeeva's scaled complex error function of complex arguments.

\[w(z) = e^{-z^2} \text{erfc}(-iz) = \text{erfcx}(-iz)\]

Examples

julia> [ (x=x, w=ComplexF16(faddeeva_w(complex(x)))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, w::ComplexF16}}:
 (x = -2.0, w = Float16(0.01831) - Float16(0.34)im)
 (x = -1.5, w = Float16(0.1054) - Float16(0.4832)im)
 (x = -1.0, w = Float16(0.368) - Float16(0.607)im)
 (x = -0.5, w = Float16(0.779) - Float16(0.479)im)
 (x = 0.0, w = Float16(1.0) + Float16(0.0)im)
 (x = 0.5, w = Float16(0.779) + Float16(0.479)im)
 (x = 1.0, w = Float16(0.368) + Float16(0.607)im)
 (x = 1.5, w = Float16(0.1054) + Float16(0.4832)im)
 (x = 2.0, w = Float16(0.01831) + Float16(0.34)im)

julia> [ (x=x, w=faddeeva_w(complex(x))) for x in (-Inf, -0.0, 0.0, Inf) ]
4-element Vector{@NamedTuple{x::Float64, w::ComplexF64}}:
 (x = -Inf, w = 0.0 - 0.0im)
 (x = -0.0, w = 1.0 - 0.0im)
 (x = 0.0, w = 1.0 + 0.0im)
 (x = Inf, w = 0.0 + 0.0im)

julia> faddeeva_w(complex(pi)) - erfcx(-im * complex(pi))
0.0 + 0.0im

See also: erfc(z), erfcx(z)

References

DLMF: 7.2.3

Faddeeva function - Wikipedia

using UnicodePlots
re_w(x) = real(LibCerf.re_w(x, 0.0))
lineplot(-4, 4, [re_w, im_w];
    title="Re[w(z)] and Im[w(z)]", ylabel="w(z)", ylim=(1, -0.75))

              ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[w(z)] and Im[w(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀        
              ┌────────────────────────────────────────┐        
            1 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠋⡏⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_w(x)
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢰⠃⠀⡇⠈⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ im_w(x)
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠇⠀⠀⡇⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠀⠀⠀⡇⠀⠀⣵⠒⠢⣄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠁⠀⠀⠀⡇⠀⡜⠀⢇⠀⠀⠣⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⡇⡰⠁⠀⠈⡆⠀⠀⠈⠣⣄⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠇⠀⠀⠀⠀⠀⣇⠇⠀⠀⠀⠘⡄⠀⠀⠀⠀⠉⠒⠤⣄⣀⡀⠀⠀⠀│        
   w(z)       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠃⠀⠀⠀⠀⠀⠀⡟⠀⠀⠀⠀⠀⠈⢆⡀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠉⠓│        
              │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠶⠶⠭⠤⠤⠤⠤⠤⠤⠤⢤⡧⠤⠤⠤⠤⠤⠤⠤⠭⠷⠶⠤⠤⠤⠤⠤⠤⠤⠤⠤│        
              │⣀⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠉⠉⠑⠒⠢⢄⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠈⠑⢆⡀⠀⠀⠀⠀⠀⠀⡎⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⢆⡀⠀⠀⠀⡸⠁⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⠦⣀⠔⠁⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
        -0.75 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
              └────────────────────────────────────────┘        
              ⠀-4⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀4⠀        
              ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀        

LibCerf.im_w — Function

im_w(x)
im_w(re, im)

Imaginary part of Faddeeva's scaled complex error function of real arguments.

im_w(x) = imag( faddeeva_w(complex(x, 0.0)) )

im_w(re, im) = imag( faddeeva_w(complex(re, im)) )

Examples

julia> [ (x=x, w=Float32(im_w(x, 0.0))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, w::Float32}}:
 (x = -2.0, w = -0.34002623)
 (x = -1.5, w = -0.48322734)
 (x = -1.0, w = -0.6071577)
 (x = -0.5, w = -0.47892517)
 (x = 0.0, w = 0.0)
 (x = 0.5, w = 0.47892517)
 (x = 1.0, w = 0.6071577)
 (x = 1.5, w = 0.48322734)
 (x = 2.0, w = 0.34002623)

julia> [ (x=x, w_im=im_w(x)) for x in (-Inf, -0.0, 0.0, Inf) ]
4-element Vector{@NamedTuple{x::Float64, w_im::Float64}}:
 (x = -Inf, w_im = -0.0)
 (x = -0.0, w_im = -0.0)
 (x = 0.0, w_im = 0.0)
 (x = Inf, w_im = 0.0)

julia> faddeeva_w(3.14+0im) - complex(re_w(3.14,0.0), im_w(3.14))
0.0 + 0.0im

See also: faddeeva_w(z), re_w(re, im)

using UnicodePlots
lineplot(-5, 5, im_w; title="Im[w(z)]", ylabel="im_w(x)", ylim=(0.75, -0.75))

                 ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Im[w(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀        
                 ┌────────────────────────────────────────┐        
            0.75 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ im_w(x)
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⢠⠖⠦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⢀⠇⠀⠀⠱⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⡸⠀⠀⠀⠀⠈⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣇⡇⠀⠀⠀⠀⠀⠀⠓⢄⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣿⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠒⠦⢄⣀⡀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡏⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⠑⠒│        
   im_w(x)       │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⢤⡧⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│        
                 │⠤⢄⣀⣀⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠈⠉⠑⠒⠤⣄⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⠦⡀⠀⠀⠀⠀⠀⢸⠁⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⡄⠀⠀⠀⠀⡎⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠘⢆⠀⠀⢠⠃⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠓⠤⠃⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
           -0.75 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                 └────────────────────────────────────────┘        
                 ⠀-5⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀5⠀        
                 ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀        

LibCerf.re_w — Function

re_w(re, im)

Real part of Faddeeva's scaled complex error function of real arguments.

re_w(re, im) = real( faddeeva_w(complex(re, im)) )

Examples

julia> [ (x=x, re_w=Float32(re_w(x, 0.0))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, re_w::Float32}}:
 (x = -2.0, re_w = 0.01831564)
 (x = -1.5, re_w = 0.10539922)
 (x = -1.0, re_w = 0.36787945)
 (x = -0.5, re_w = 0.7788008)
 (x = 0.0, re_w = 1.0)
 (x = 0.5, re_w = 0.7788008)
 (x = 1.0, re_w = 0.36787945)
 (x = 1.5, re_w = 0.10539922)
 (x = 2.0, re_w = 0.01831564)

julia> [ (x=x, re_w=re_w(x, 0.0)) for x in (-Inf, 0.0, Inf) ]
3-element Vector{@NamedTuple{x::Float64, re_w::Float64}}:
 (x = -Inf, re_w = 0.0)
 (x = 0.0, re_w = 1.0)
 (x = Inf, re_w = 0.0)

julia> faddeeva_w(1.0+2.0im) - complex(re_w(1.0,2.0), im_w(1.0,2.0))
0.0 + 0.0im

See also: faddeeva_w(z), im_w(re, im)

using UnicodePlots
re_w(x) = real(LibCerf.re_w(x, 0.0))
lineplot(-2.5, 2.5, re_w; title="Re[w(z)]", ylabel="re_w(x, 0.0)")

                  ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[w(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀        
                  ┌────────────────────────────────────────┐        
                1 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠉⡏⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_w(x)
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢰⠁⠀⡇⠈⣆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠇⠀⠀⡇⠀⠸⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡸⠀⠀⠀⡇⠀⠀⢇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⡇⠀⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠀⠀⠀⠀⡇⠀⠀⠀⢣⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢰⠃⠀⠀⠀⠀⡇⠀⠀⠀⠘⡆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
   re_w(x, 0.0)   │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡏⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣸⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⢇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠘⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                  │⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠈⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀│        
                0 │⠀⠀⠀⣀⣀⣀⡠⠴⠊⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⠢⢄⣀⣀⡀⠀⠀⠀│        
                  └────────────────────────────────────────┘        
                  ⠀-3⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀3⠀        
                  ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀        

Dawson’s Integral

LibCerf.dawson — Function

dawson(x)
dawson(z)

Dawson integral of real or complex arguments.

\[\text{Dawson}(z) = e^{-z^2} \int_0^z e^{t^2} dt = \frac{\sqrt{\pi}}{2} e^{-z^2} \text{erfi}(z)\]

Examples

julia> [ (x=x, dawson=Float32(dawson(x))) for x in -2:0.5:2 ]
9-element Vector{@NamedTuple{x::Float64, dawson::Float32}}:
 (x = -2.0, dawson = -0.3013404)
 (x = -1.5, dawson = -0.42824906)
 (x = -1.0, dawson = -0.5380795)
 (x = -0.5, dawson = -0.4244364)
 (x = 0.0, dawson = 0.0)
 (x = 0.5, dawson = 0.4244364)
 (x = 1.0, dawson = 0.5380795)
 (x = 1.5, dawson = 0.42824906)
 (x = 2.0, dawson = 0.3013404)

julia> [ (x=x, dawson=dawson(x)) for x in (-Inf, -0.0, 0.0, Inf) ]
4-element Vector{@NamedTuple{x::Float64, dawson::Float64}}:
 (x = -Inf, dawson = -0.0)
 (x = -0.0, dawson = -0.0)
 (x = 0.0, dawson = 0.0)
 (x = Inf, dawson = 0.0)

julia> abs( dawson(complex(pi)) - sqrt(pi)/2*exp(-pi^2)*erfi(complex(pi)) ) < eps(Float64)
true

See also: erfi(z)

References

DLMF: 7.2.5

Dawson's Integral – Wolfram MathWorld

Real part plot of dawson(z):

using UnicodePlots
re_dawson(x) = real(LibCerf.dawson(complex(x)))
lineplot(-5, 5, re_dawson;
    title="Re[Dawson(z)]", ylabel="dawson(x+0.0im)", ylim=(0.75, -0.75))

                         ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Re[Dawson(z)]⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀             
                         ┌────────────────────────────────────────┐             
                    0.75 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ re_dawson(x)
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⡰⠋⠑⣄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⢰⠃⠀⠀⠀⠣⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⡇⠀⠀⠀⠀⠀⠑⢆⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣿⠁⠀⠀⠀⠀⠀⠀⠀⠉⠒⠤⣄⣀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡏⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠑⠒⠒⠒│             
   dawson(x+0.0im)       │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⢤⡧⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│             
                         │⠤⠤⠤⢄⣀⣀⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡸⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠉⠉⠒⠦⣄⠀⠀⠀⠀⠀⠀⠀⠀⡇⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⢄⠀⠀⠀⠀⠀⡸⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠣⡀⠀⠀⢀⠇⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⢆⣀⠜⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                   -0.75 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│             
                         └────────────────────────────────────────┘             
                         ⠀-5⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀5⠀             
                         ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀             

Voigt Functions

LibCerf.voigt — Function

voigt(x, sigma, gamma)

Convolution of a Gaussian and a Lorentzian of real arguments.

\[\text{Voigt}(x; \sigma, \gamma) = \int_{-\infty}^{\infty} G(t; \sigma) L(x-t; \gamma) dt\]

where

\[G(x; \sigma) = \frac{1}{\sigma\sqrt{2\pi}} \exp(-\frac{x^2}{2\sigma^2})\]

\[L(x; \gamma) = \frac{ \gamma }{ \pi(x^2+\gamma^2) }\]

Examples

julia> [ (x=x, voigt=Float32(voigt(x, 1.0, 0.0))) for x in -10:2.0:10 ]
11-element Vector{@NamedTuple{x::Float64, voigt::Float32}}:
 (x = -10.0, voigt = 7.6945985f-23)
 (x = -8.0, voigt = 5.052271f-15)
 (x = -6.0, voigt = 6.075883f-9)
 (x = -4.0, voigt = 0.00013383022)
 (x = -2.0, voigt = 0.053990968)
 (x = 0.0, voigt = 0.3989423)
 (x = 2.0, voigt = 0.053990968)
 (x = 4.0, voigt = 0.00013383022)
 (x = 6.0, voigt = 6.075883f-9)
 (x = 8.0, voigt = 5.052271f-15)
 (x = 10.0, voigt = 7.6945985f-23)

julia> [ (x=x, voigt=Float32(voigt(x, 0.0, 3.0))) for x in -10:2.0:10 ]
11-element Vector{@NamedTuple{x::Float64, voigt::Float32}}:
 (x = -10.0, voigt = 0.008760823)
 (x = -8.0, voigt = 0.013081228)
 (x = -6.0, voigt = 0.02122066)
 (x = -4.0, voigt = 0.038197186)
 (x = -2.0, voigt = 0.07345613)
 (x = 0.0, voigt = 0.10610329)
 (x = 2.0, voigt = 0.07345613)
 (x = 4.0, voigt = 0.038197186)
 (x = 6.0, voigt = 0.02122066)
 (x = 8.0, voigt = 0.013081228)
 (x = 10.0, voigt = 0.008760823)

julia> [ (x=x, voigt=Float32(voigt(x, 1.0, 0.5))) for x in (-Inf, 0.0, Inf) ]
3-element Vector{@NamedTuple{x::Float64, voigt::Float32}}:
 (x = -Inf, voigt = 0.0)
 (x = 0.0, voigt = 0.27895546)
 (x = Inf, voigt = 0.0)

See also: faddeeva_w(z), voigt_hwhm(sigma, gamma)

References

Voigt profile - Wikipedia

using UnicodePlots
# Plot Voigt function
σ, γ = 2.0, 1.0;
gauss(x) = LibCerf.voigt(x, σ, 0.0)     # Gauss (Normal) distribution
lorentz(x) = LibCerf.voigt(x, 0.0, γ)   # Cauchy–Lorentz distribution
voigt_σγ(x) = LibCerf.voigt(x, σ, γ)
plt = lineplot(-5, 5, [gauss, lorentz, voigt_σγ]
    ; title="Voigt(x; σ, γ), G(x; σ), L(x; γ)"
    , height=20, width=60)

            ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Voigt(x; σ, γ), G(x; σ), L(x; γ)⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀            
            ┌────────────────────────────────────────────────────────────┐            
        0.2 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⣀⠔⠋⠉⡏⠉⠢⣄⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ gauss(x)   
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⠝⠀⠀⠀⠀⡇⠀⠀⠀⠫⡆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ lorentz(x) 
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⡻⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢫⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ voigt_σγ(x)
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠁⡇⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⢸⠈⢣⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡜⠁⢸⠁⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⡇⠀⢣⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡜⠀⠀⡜⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⢱⠀⠀⢣⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡜⠀⠀⢠⠇⢀⡤⠒⠊⠉⠉⡏⠉⠑⠒⢤⡀⠸⡄⠀⠀⢱⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡜⠀⠀⠀⡼⠖⠁⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠈⠓⢧⡀⠀⠀⢱⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡜⠀⠀⡠⢪⠇⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠘⡕⢆⡀⠀⢱⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
   f(x)     │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡜⠀⡠⠊⠀⡎⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⢣⠀⠑⢆⡀⢱⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡜⡠⠎⠀⠀⢰⠁⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠈⢇⠀⠀⠱⢄⢱⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣎⠖⠁⠀⠀⢠⠇⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠘⡄⠀⠀⠀⠓⣵⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠞⠁⠀⠀⠀⢀⠎⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠸⡄⠀⠀⠀⠀⠳⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⢴⠏⠀⠀⠀⠀⢠⠎⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠱⡄⠀⠀⠀⠀⠸⡧⡄⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⠀⠀⠀⣀⠔⡡⠃⠀⠀⠀⠀⡠⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠘⢄⠀⠀⠀⠀⠘⢎⠣⣄⠀⠀⠀⠀⠀⠀⠀⠀│            
            │⠀⠀⠀⠀⠀⢀⡠⠊⢁⠖⠁⠀⠀⠀⣀⠖⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠓⣄⠀⠀⠀⠀⠓⡌⠑⢄⡀⠀⠀⠀⠀⠀│            
            │⠀⠀⢀⡠⠞⠁⢀⠔⠃⠀⠀⢀⡴⠊⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠑⠦⡀⠀⠀⠘⠢⡀⠈⠳⢄⡀⠀⠀│            
            │⠔⠊⠁⠀⣠⠔⢃⣀⡤⠔⠊⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠑⠢⢤⣀⡈⠢⣄⠀⠈⠑⠢│            
            │⣤⣤⠤⠞⠓⠉⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠚⠓⠦⣤⡤│            
          0 │⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈│            
            └────────────────────────────────────────────────────────────┘            
            ⠀-5⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀5⠀            
            ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀            

LibCerf.voigt_hwhm — Function

voigt_hwhm(sigma, gamma)

Half width at half maximum of the Voigt function of real arguments.

It is implicitly defined by:

\[\text{Voigt}( \text{VoigtHWHM}( \sigma, \gamma); \sigma, \gamma) = \frac{1}{2} \text{Voigt}(0; \sigma, \gamma)\]

Examples

julia> σ, γ = 1.3, 0.5;

julia> [ (x=x, voigt=Float32(voigt(x, σ, γ))) for x in -12:4.0:12 ]
7-element Vector{@NamedTuple{x::Float64, voigt::Float32}}:
 (x = -12.0, voigt = 0.0011444782)
 (x = -8.0, voigt = 0.0027030057)
 (x = -4.0, voigt = 0.016749112)
 (x = 0.0, voigt = 0.23147874)
 (x = 4.0, voigt = 0.016749112)
 (x = 8.0, voigt = 0.0027030057)
 (x = 12.0, voigt = 0.0011444782)

julia> Float32(voigt(0.0, σ, γ))    # peak
0.23147874f0

julia> Float32(voigt(0.0, σ, γ)) / 2  # half peak
0.11573937f0

julia> hwhm_x = voigt_hwhm(σ, γ);   # here voigt takes exactly half the peak value

julia> voigt(hwhm_x, σ, γ) ≈ voigt(0.0, σ, γ)/2
true

See also: voigt(x, sigma, gamma)

References

Wuttke, J. (2025). Power series for the half width of the Voigt function, rederived. Journal of Numerical Analysis and Approximation Theory, 54(2), 345-356. https://doi.org/10.33993/jnaat542-1640

using UnicodePlots
# Plot Voigt function
σ, γ = 1.3, 0.5;
voigt_σγ(x) = LibCerf.voigt(x, σ, γ)
plt = lineplot(-5, 5, voigt_σγ; title="Voigt(x; σ=$σ, γ=$γ)", ylim=(0.25, -0.05))
# Plot y_half and x_hwhm
y_peak = LibCerf.voigt(0.0, σ, γ)
y_half = Float32(y_peak / 2)        # Half peak
x_hwhm = Float32(LibCerf.voigt_hwhm(σ, γ))  # Get x where y = y_half
hline!(plt, y_half; name="y_half = $y_half")
vline!(plt, x_hwhm; name="x_hwhm = $x_hwhm")

              ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Voigt(x; σ=1.3, γ=0.5)⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀                    
              ┌────────────────────────────────────────┐                    
         0.25 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣇⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ voigt_σγ(x)        
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⠎⠁⡇⠑⣄⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ y_half = 0.11573937
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠃⠀⠀⡇⠀⠈⢆⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ x_hwhm = 1.815095  
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠁⠀⠀⠀⡇⠀⠀⠈⢇⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠁⠀⠀⠀⠀⡇⠀⠀⠀⠈⢇⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢰⠃⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠈⢆⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              │⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⢤⠧⠤⠤⠤⠤⠤⠤⡧⠤⠤⠤⠤⠤⠼⡧⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤⠤│                    
   f(x)       │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡟⡆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠃⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡇⠘⢆⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⡔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡇⠀⠈⠣⡀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              │⠀⠀⠀⠀⠀⠀⢀⡤⠊⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠑⠦⡀⠀⠀⠀⠀⠀⠀│                    
              │⠀⣀⣀⡠⠤⠒⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠈⠓⠦⢄⣀⣀⠀│                    
              │⠛⠓⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⡗⠒⠒⠒⠒⠒⠒⡗⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠒⠛│                    
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
        -0.05 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│                    
              └────────────────────────────────────────┘                    
              ⠀-5⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀5⠀                    
              ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀x⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀